% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeT2.R
\name{computeT2}
\alias{computeT2}
\title{Compute T^2 and its p-value for each pathway}
\usage{
computeT2(data, purb = 1.5, pathDB = "KEGG", ppi = STRING_v91,
  intg = TRUE, alpha = 0.05, ncore = 7)
}
\arguments{
\item{data}{Processed data using importdata function.}

\item{purb}{Perturbance threshold, default is 1.5 (after normalization).}

\item{pathDB}{Pathway database: "KEGG" or "Reactome".}

\item{ppi}{Protein-protein interaction database: "STRING_ppi" or "HitPredict_ppi".}

\item{intg}{Apply pathway integration or not, default is TRUE.}

\item{alpha}{Significance level, default is 0.05.}

\item{ncore}{Number of parallel computing cores, default is 7.}
}
\description{
This function computes the T^2 score and its significance level.
}
\examples{
dat1=importdata(TCR_5min)
res1=computeT2(dat1)
dat2=importdata(TCR_5min,TCR_15min)
res2=computeT2(dat2,pathDB="Reactome",ppi=HitPredict_v4)
}
\keyword{$T^2$}
\keyword{compute}
