% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeT2.R
\name{computeT2}
\alias{computeT2}
\title{Compute T^2 and its p-value for each pathway}
\usage{
computeT2(data, purb = 1.5, pathDB = "KEGG", ppi = STRING_ppi,
  intg = TRUE, alpha = 0.05, ncore = 1, per = 10000)
}
\arguments{
\item{data}{Processed data using importdata function.}

\item{purb}{Perturbance threshold, default is 1.5 (after normalization).}

\item{pathDB}{Pathway database: "KEGG" or "Reactome".}

\item{ppi}{Protein-protein interaction database: "STRING_ppi" or "HitPredict_ppi".}

\item{intg}{Apply pathway integration or not, default is TRUE.}

\item{alpha}{Significance level, default is 0.05.}

\item{ncore}{Number of parallel computing cores, default is 7.}

\item{per}{Number of permutation to construct the null distribution.}
}
\description{
This function computes the T^2 score and its significance level.
}
\examples{
dat1=importdata(tcr_05)
dat2=importdata(tcr_05,tcr_15)
res1=computeT2(dat1)
res2=computeT2(dat2,vex=Reactome_vex,ppi=HitPredict_ppi)
}
\keyword{$T^2$}
\keyword{compute}

